+++ README.md
@@ -1,70 +1,44 @@
-# Getting Started with Create React App
+# 폴더 구조
 
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+## 📁 public
 
-## Available Scripts
+- index.html
 
-In the project directory, you can run:
+## 📁 src
 
-### `npm start`
+1. assets
 
-Runs the app in the development mode.\
-Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
+- image, video, json
 
-The page will reload when you make changes.\
-You may also see any lint errors in the console.
+2. components
 
-### `npm test`
+- 공통 컴포넌트
+  - Header, Footer, Nav, Button ...
 
-Launches the test runner in the interactive watch mode.\
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
+3. pages
 
-### `npm run build`
+- 페이지 단위의 컴포넌트
+  - userlistPage, adminlist, strlist
+    - userlistContainer -> userlistheader, userllistBody, userlistFooter
+    - userlistheader, table, search form
 
-Builds the app for production to the `build` folder.\
-It correctly bundles React in production mode and optimizes the build for the best performance.
+## 📁 styles
 
-The build is minified and the filenames include the hashes.\
-Your app is ready to be deployed!
+- `styled-components` 공통 속성
+  - GlobalStyle, theme...
 
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
+## 📁 services
 
-### `npm run eject`
+- api 로직
 
-**Note: this is a one-way operation. Once you `eject`, you can't go back!**
+## 📁 utils
 
-If you aren't satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
+- 상수나 공통 함수, 유틸리티를 담는 폴더.
 
-Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you're on your own.
+## 📁 hooks
 
-You don't have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn't feel obligated to use this feature. However we understand that this tool wouldn't be useful if you couldn't customize it when you are ready for it.
+- 커스텀 훅
 
-## Learn More
+## 📁 store
 
-You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
-
-To learn React, check out the [React documentation](https://reactjs.org/).
-
-### Code Splitting
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/code-splitting](https://facebook.github.io/create-react-app/docs/code-splitting)
-
-### Analyzing the Bundle Size
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size](https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size)
-
-### Making a Progressive Web App
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app](https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app)
-
-### Advanced Configuration
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/advanced-configuration](https://facebook.github.io/create-react-app/docs/advanced-configuration)
-
-### Deployment
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/deployment](https://facebook.github.io/create-react-app/docs/deployment)
-
-### `npm run build` fails to minify
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify](https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify)
+- redux 로직
