+++ src/App.js
@@ -3,7 +3,6 @@ import { Routes, Route } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
 import theme from './theme';
 import Nav from './components/common/Nav';
-import Example from './components/common/Example';
 import Home from './components/common/Home';
 import Login from './components/login/LoginComponent';
 import UserList from './components/user/UserListComponent';
@@ -26,7 +25,6 @@ function App() {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/login" element={<Login />} />
-        <Route path="/test" element={<Example />} />
         <Route path="/user" element={<UserList />} />
         <Route path="/user/:no" element={<UserDetail />} />
         <Route path="/stretching" element={<StretchingList />} />
