+++ src/components/common/Winner.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Main from './Main';
 import Content from './Content';
-import Button from './elements/Button';
+import Button from './button/Button';
 
 function Winner() {
   return (
