+++ src/components/event/RegisterWinner.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import Button from '../common/elements/Button';
+import Button from '../common/button/Button';
 import Content from '../common/Content';
 import InputBox from '../common/elements/InputBox';
 import Main from '../common/Main';
