+++ src/components/stretching/RegisterStretching.js
@@ -5,7 +5,7 @@ import { changeField, clearField } from '../../store/stretching/wirteSlice';
 
 import Main from '../common/Main';
 import Content from '../common/Content';
-import Button from '../common/elements/Button';
+import Button from '../common/button/Button';
 import Editor from '../editor/Editor';
 
 // useMemo : 컴포넌트 최적화
