+++ src/pages/user/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import UserList from '../../components/user/UserListComponent';
+
+function User() {
+  return <UserList />;
+}
+
+export default User;
